feat: skip size json generation for targets in preview
Targets in preview may not yet have support in esp-idf-size, which
may cause a failure while the size json file is generated. This may block
chip support team as they need to wait or work on the esp-idf-size
support first.

This skips size json generation if target is in preview.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>